+++ include/median_calculator.h
@@ -2,7 +2,7 @@
 #define MEDIAN_CALCULATOR_H
 class MedianCalculator {
   public:
-    virtual float GetMedian() = 0;
+    virtual double GetMedian() = 0;
     virtual void AddValue(int value) = 0;
 };
 #endif
