+++ src/array_median.cpp
@@ -1,6 +1,6 @@
 #include "array_median.h"
 
-float ArrayMedianCalculator::GetMedian(void) {
+double ArrayMedianCalculator::GetMedian(void) {
   if (size % 2 == 0) {
     return ((long long int)values[(size/2)-1] + values[size/2])/2.0;
   } else {
